+++ Assignment-03/pc_vs2/SmartRoom/getC2Dms.h
@@ -21,7 +21,7 @@ public:
     IOTHUB_DEVICE_CLIENT_LL_HANDLE endService();
     //IOTHUB_DEVICE_CLIENT_LL_HANDLE sync();
     bool getDetectStatus();
-    int getLightValue();
+    double getLightValue();
     bool getDashboardSwitch();
     bool getLightSwitchNode();
     int getRollerSliderValueNode();
